Fix #17 removed lock dependencies
This library has no dependencies at all, but only runs in Chrome extensions.

However, for #8 we added dev dependencies, which are _only_ used when compiling legacy ES5 shims for React. This is an extreme corner case as the existing Chrome API works fine with ES5, and this library only adds value for people who wish to use `Promise` (ES2016) or `await` (ES2017).

One of the tools used for the ES5 legacy compilation is Babel, which has a lot of NPM dependencies, 2 of which have security alerts that we don't want to pass on.

Deleting the lock file means you cannot reproduce the exact versions of dev tools used, but as that doesn't affect production we recommend that you use the very latest versions, should you want ES5 support for ES2017-only features.