Handle edge case with no callbackArgs
This is an alternative approach to fixing #23.  For chrome functions that pass no arguments to their callbacks, we should try to propagate the function's own return value.

Honestly, I'm pretty confused about how it is even able to work.  It makes sense, kind of, but I'm surprised that I can access the function return inside an arrow function defined as an inline function argument.  In any case, I tested and confirmed that this works.

Fixes #23 and closes #26.